Initial commit of clean project interface for moving forward
Removed all legacy code for detection and tracking (bbc project) and added new interface and message design.
Very much a WIP atm.
