--- config/bbc_detectors.yaml
@@ -1,17 +1,0 @@
-bayes_people_tracker:
-    filter_type: "UKF"                                         # The Kalman filter type: EKF = Extended Kalman Filter, UKF = Uncented Kalman Filter
-    prune_named: false                                           # make sure identity tracks are unique (prunes the one with high stdev)
-    cv_noise_params:                                           # The noise for the constant velocity prediction model
-        x: 0.05
-        y: 0.05
-    std_limit: 5.0                                              # pruning tracks that have std(X) > std_limit
-    detectors:                                                 # Add detectors under this namespace
-        ripe_strawberry_detector:                                   # Name of detector (used internally to identify them). Has to be unique.
-            topic: "/detection/pose_array"                # The topic on which the geometry_msgs/PoseArray is published
-            cartesian_noise_params:                            # The noise for the cartesian observation model
-                x: 0.025
-                y: 0.025
-            seq_size: 1                                        # minimum `seq_size` number of detections received in `seq_time` seconds
-            seq_time: 0.25                                      #   to create new track
-            observation_model: "CARTESIAN"
-            matching_algorithm: "NN"                       # The algorithm to match different detections. NN = Nearest Neighbour, NNJPDA = NN Joint Probability Data Association
