--- config/detectors.yaml
@@ -1,13 +1,0 @@
-bayes_people_tracker:
-    filter_type: "UKF"                                         # The Kalman filter type: EKF = Extended Kalman Filter, UKF = Uncented Kalman Filter
-    cv_noise_params:                                           # The noise for the constant velocity prediction model
-        x: 1.4
-        y: 1.4
-    detectors:                                                 # Add detectors under this namespace
-        strawberry_detector:                                   # Name of detector (used internally to identify them). Has to be unique.
-            topic: "/rasberry_perception/detection/pose_array" # The topic on which the geometry_msgs/PoseArray is published
-            cartesian_noise_params:                            # The noise for the cartesian observation model
-                x: 0.1
-                y: 0.1
-            observation_model: "CARTESIAN"
-            matching_algorithm: "NNJPDA"                       # The algorithm to match different detections. NN = Nearest Neighbour, NNJPDA = NN Joint Probability Data Association
