--- description/d435/README.md
@@ -1,3 +1,0 @@
-# Intel RealSense D435 Description
-
-Visualisation files for RVIZ taken from the [realsense-ros](https://github.com/IntelRealSense/realsense-ros/tree/c2448916218ccfe49b0d642563493cb4e9bdcc3b/realsense2_description) repo.
