+++ msg/Detections.msg
@@ -1,0 +1,6 @@
+# This message is for storing detections for an image
+#  Raymond Kirk (Tunstill) Copyright (c) 2020
+#  Email: ray.tunstill@gmail.com
+
+Header header # Detection origin header i.e. the rgb header the detections are from
+rasberry_perception/Detection[] detections # Detection results for the image
