+++ msg/SegmentOfInterest.msg
@@ -1,0 +1,5 @@
+# This message is for storing a segment of interest (i.e. segmentation results) in a generic format
+#  Raymond Kirk (Tunstill) Copyright (c) 2020
+#  Email: ray.tunstill@gmail.com
+uint32[] x  # Array of [x1 x2 ... xn] in image co-ordinates
+uint32[] y  # Array of [y1 y2 ... yn] in image co-ordinates
