+++ package.xml
@@ -1,12 +1,17 @@
 <?xml version="1.0"?>
+<!--
+  ~ Raymond Kirk (Tunstill) Copyright (c) 2020
+  ~ Email: ray.tunstill@gmail.com
+  -->
+
 <package format="2">
   <name>rasberry_perception</name>
   <version>0.0.0</version>
   <description>The rasberry_perception package</description>
 
-  <maintainer email="rkirk@lincoln.ac.uk">raymond</maintainer>
+  <maintainer email="ray.tunstill@gmail.com">Raymond</maintainer>
 
-  <license>TODO</license>
+  <license>MIT (see LICENSE in project root)</license>
 
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>roscpp</build_depend>
@@ -33,5 +38,4 @@
   <exec_depend>message_runtime</exec_depend>
   <exec_depend>dynamic_reconfigure</exec_depend>
   <exec_depend>rosserial</exec_depend>
-
 </package>
