+++ scripts/examples/__init__.py
@@ -1,0 +1,3 @@
+#  Raymond Kirk (Tunstill) Copyright (c) 2020
+#  Email: ray.tunstill@gmail.com
+
