--- src/bbc_countryfile/models.py
@@ -1,22 +1,0 @@
-model_lookup = {
-    "faster-rcnn": {
-        "config": "/home/rasberry/raymond_ws/src/mmdetection/configs/bbc_countryfile/faster_rcnn_r50_fpn_1x_2cls_1333x800.py",
-        "model": "/home/rasberry/raymond_ws/src/mmdetection/work_dirs/bbc_countryfile/faster_rcnn_r50_fpn_1x_2cls_1333x800/epoch_12.pth"
-    },
-    "faster-rcnn-16": {
-        "config": "/home/rasberry/raymond_ws/src/mmdetection/configs/bbc_countryfile/faster_rcnn_r50_fpn_fp16_1x_2cls_1333x800.py",
-        "model": "/home/rasberry/raymond_ws/src/mmdetection/work_dirs/bbc_countryfile/faster_rcnn_r50_fpn_fp16_1x_2cls_1333x800/epoch_12.pth"
-    },
-    "ssd": {
-        "config": "/home/rasberry/raymond_ws/src/mmdetection/configs/bbc_countryfile/ssd512_1x_2cls_512x512.py",
-        "model": "/home/rasberry/raymond_ws/src/mmdetection/work_dirs/bbc_countryfile/ssd512_1x_2cls_512x512/epoch_12.pth"
-    },
-    "grid-rcnn": {
-        "config": "/home/rasberry/raymond_ws/src/mmdetection/configs/bbc_countryfile/grid_rcnn_gn_head_r50_fpn_1x_2cls_1333x800.py",
-        "model": "/home/rasberry/raymond_ws/src/mmdetection/work_dirs/bbc_countryfile/grid_rcnn_gn_head_r50_fpn_1x_2cls_1333x800/epoch_12.pth"
-    },
-    "reppoints": {
-        "config": "/home/rasberry/raymond_ws/src/mmdetection/configs/bbc_countryfile/reppoints_moment_r50_fpn_2x.py",
-        "model": "/home/rasberry/raymond_ws/src/mmdetection/work_dirs/bbc_countryfile/reppoints_moment_r50_fpn_2x/epoch_24.pth"
-    },
-}
