+++ src/rasberry_perception/detection_interface/compatibility_layer/catalog.py
@@ -1,3 +1,4 @@
 #  Raymond Kirk (Tunstill) Copyright (c) 2020
 #  Email: ray.tunstill@gmail.com
 
+# PLACEHOLDER FILE FOR DOWNLOADING MODELS
