+++ src/rasberry_perception/detection_interface/compatibility_layer/python3_fixes.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python
+
+#  Raymond Kirk (Tunstill) Copyright (c) 2020
+#  Email: ray.tunstill@gmail.com
+
+from __future__ import absolute_import, division, print_function
+import sys
+
+
+class KineticImportsFix:
+    def __init__(self, kinetic_dist_packages="/opt/ros/kinetic/lib/python2.7/dist-packages"):
+        self.kinetic_dist_packages = kinetic_dist_packages
+
+    def __enter__(self):
+        sys.path.remove(self.kinetic_dist_packages)
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        sys.path.append(self.kinetic_dist_packages)
