+++ srv/GetDetectorResults.srv
@@ -1,5 +1,6 @@
+# Service for DeepLearning backend interface with ROS system (sensor_msgs/Image, threshold_param) -> detections
 sensor_msgs/Image image
 float32 score_thresh
 ---
-string status
-rasberry_perception/ImageDetections detections
+rasberry_perception/DetectionStatus status
+rasberry_perception/Detections detections
